requirements.txt
streamlit>=1.20
opencv-python-headless>=4.7.0
mediapipe>=0.10.0
numpy>=1.24
Pillow>=9.5
